Fix conversion warnings, add move ls, and set a generic rotating heuristic

+++ src/data/models/Solution.hpp
@@ -19,7 +19,7 @@ namespace routing {
                 virtual double getCost() = 0;
                 virtual void update() = 0;
                 virtual void pushTour(Tour* tour) = 0;
-                virtual void addTour(Tour* tour, unsigned position) = 0;
+                virtual void addTour(Tour* tour, unsigned long position) = 0;
                 virtual unsigned long getNbTour() const = 0;
                 virtual void getVarsVals(IloNumVarArray & vars, IloNumArray & vals) = 0;
                 virtual void print(std::ostream & out) = 0;
